graphql-composition: generic @link support (#2565)
Composition previously only dealt with `@link` for the federation
directives. Now, we take into account every `@link` and internalize them
in the Subgraphs struct. The federated graph output is not changed at
the moment, because only the only linked schema with an effect is the
federation schema. But this is the infrastructure we need for arbitrary
directives from extensions.

closes GB-8313