graphql-composition: use insta for tests (#2573)
Each test root is identified by a `test.md` file which can have a
description of the purpose of the test. The contents of the test.md will
end up in the snapshot for the federated graph, directly next to the
diff, which I find pretty neat for documenting what the test does when
you are reviewing changes to the snapshot.