feat: loader/host side of authentication extensions (#2576)
This adds the wasi loader side of auth extensions with a concurrent
cache. I'm doing the actual authenticator for JWT and running the
integration tests in a subsequent pull request.

## Cache

We can't have all extensions triggering HTTP requests concurrently, e.g.
when fetching JWKs for a JWT authenticator. What we do here is we let
the first one getting a cache miss from the host, letting the guest to
run the init function to get the needed value. Other callers will now
wait for this one fetcher to be done, and they will get the value from
the init function when it's finished.

## Configuration

The auth extensions need special configuration. We need to figure out
how to pass this to the loader. We will need a separate loader per
authenticator, so if the user has more than one auth provider with the
same extension, they will run in separate instances. The current idea
defined in the RFC is to let the authentication config define
configuration for the extension, I'm not completely sure how great this
idea is, but we'll see again in the next pull request how it's going to
look like.