graphql-composition: fix @link aliased directive imports, and enforce import types (#2603)
Take the `@` prefix into account in aliased imports from `@link`
directives. Previously, the `as:` argument for individual directives
would only work if you did not prefix the alias with an `@`. But the
link spec is clear: if you import a directive, you have to import it
with an `@` prefix, and other types should be imported without. This
behaviour is fixed, and the presence or absence of `@` is now enforced.

By import types, "directive" or "not directive" is meant here.

closes GB-8446