feat(engine): Faster response deserialization
One of the parts that took significant time in deserialization was the
binary search to detect which field we're currently reading from the
subgraph response as we don't rely on any particular ordering.

I've replaced with a linear search with a increasing offset if the
subgraph response does come in the expected order (for a GraphQL query
at least). It's ~10% faster on big subgraph responses.

We shouldn't have any negative effects from switching to a linear search
as having more than a few dozens of fields is already very unlikely.