cli: grafbase compose command (#2829)
This is long overdue. The options and the logic to fetch subgraph
schemas are taken from `grafbase dev`. If `--graph-ref` is passed, the
subgraphs from the given remote graph are taken into account, in
addition to the subgraphs from the local configuration toml (if
present). One of `--graph-ref` or the local paths must be present.

closes GB-6934