Snowflake extension init (#2825)
See the README for details. It has limitations.

One limitation we should lift before we merge this PR is taking the env
vars for the private and public key from the configuration instead of
hardcoded environment variables.

Additionally, with a bit more work, we could returned rows with named
columns instead of `[[JSON!]!]` from the resolver.

The tests mock the response instead of querying against the live
snowflake API.

closes GB-8456