chore(wasi-component-loader): Move config handling into wasi-component-loader
Today the config processing is done inside federated-server directly,
but that makes it difficult to re-use it in integration-tests and we
have better separation of concerns now imho.

This PR also adds a dummy extension in integration-tests I intend to add
use later until I realized I needed to rework the wasm extensions initialization.