gateway: configurable GDN poll url (#2990)
This is for enterprise platform users who do not use the GDN (Graph
Delivery Network), but instead their own HTTP endpoint to serve schemas.

The feature takes the form of a new gateway configuration option:

```
[graph]
schema-fetch-url = "https://my-custom-s3-bucket/my-account-id/{{ graph-ref.graph }}/{{ graph-ref.branch }}"
```

The following variables are available in the url template:

- `graph-ref.graph` : the graph name from the graph ref the gateway was
started with.
- `graph-ref.branch`: the optional branch name from the graph ref the
gateway was started with.

closes GB-8863