chore(engine): Better operation errors
Until now we were stopping at the first error. But for MCP we want to
return as much information as possible and it's useful for clients anyway.

Additionally I'm now tracking relevant `DirectiveSiteId`s so that I can
later send the relevant SDL to the LLMs based on the errors. So instead
of just saying "Oops this field doesn't exist!" I'll include the type's
GraphQL definition.