Protobuf compiler plugin to generate gRPC virtual subgraphs (#2925)
For use in combination with the upcoming grpc extension.

One invocation corresponds to the schema of one virtual GraphQL
subgraph, however many protocol buffer IDL files are given as input.

In that GraphQL schema, we have:

- A translation of the definitions in the protobuf IDL file in
`@protoMessages`, `@protoEnums` and `@protoServices` directives on the
schema definition.
- Root mutation and subscription fields (we should also support query
fields via a protobuf option soon) corresponding to service methods,
with an `@grpcMethod` directive.
- The input and output type definitions required for these methods.

Protocol buffer definitions tend to use a lot of namespacing, but that
is not a thing in GraphQL schemas. Instead, here I followed a convention
of replacing dots with underscores in fully qualified paths to use as
type names. That should avoid collisions and be relatively easy to
discover.

closes GB-8846