cli: integrate new assets for grafbase dev
The main difference with the previous iteration of the app is in the data requirements. The new app expects an `/app/data.json` endpoint that contains

I am stopping here in the interest of time, but there are weaknesses we should address in the future. In particular, `data.json` is polled, and it can be large. We should use SSE or similar here to avoid unnecessary redownloads. Or at least e-tag based caching.

closes GB-8903