chore(engine): Decouple shape id from the subgraph response
Today the `ResponsePart`, which as the name suggests is a part of the
response written by a resolver, is tied to a specific shape (computed
from the GraphQL query). This doesn't work well with `@lookup` as we add
a root "virtual" GraphQL field inside the plan, taken into account for the
shape computation. For some resolver like GraphQL we have to use the
complete shape. For others, like extensions, we have to ignore this
top field.

Decoupling both makes it possible to handle this more finely.

I've also renamed `InputObjectId` and other naming like this to
`ParentObjectId` which should be clearer. I'm also adding a reference to
`Schema` & `PreparedOperation` inside the `ResponsePart`.