fix(engine): Fix incorrect interprestation of resolvable: false in join__type
Julius noticed the issue when calling a lookup field directly. Somehow
the engine decided to retrieve the key fields and then to an entity join
to retrieve the rest. The problem was that we weren't interpreting
`resolvable: Boolean` correctly. It represents whether the entity is
available through Apollo `_entities` field, not whether any field exists
within a subgraph.