graphql-composition: test that directive definitions line up (#3114)
In graphql-composition tests, test that the federated graph is valid
GraphQL with graphql-schema-validation. That revealed a few directive
definitions that didn't line up with the usage, sometimes because the
usage was wrong, sometimes because the directive definition was wrong.

Also revealed a bug and a bad error message in
graphql-schema-validation.