cli: add third-party plugins support (#3133)
Like cargo and git.

For example, if you have a `grafbase-postgres` binary installed, you can
call it with `grafbase postgres`.

This PR includes:

- Executing third-party plugins when available
- Advertising them in the help text
- A `grafbase list-third-plugins` command to list them
- Handling of bad arguments, unknown third party plugins, etc.

The gateway (`grafbase-gateway`) is given special treatment and excluded
from plugins, since it is its own thing, and it would be confusing if
`grafbase gateway` worked locally, but not in an environment where the
CLI is not installed.

closes GB-8998