cli: always unpack assets in debug builds
In development, assets are not unpacked again until you bump the version in Cargo.toml. This is not convenient for iteration, so this PR makes it so that debug builds of the CLI will always unpack the assets they were compiled with, which means slower startup, but potentially easier iteration on the CLI app integration.

closes GB-9037