feat(engine): Add `@derive` directive
Instead of relying solely on `@is` we introduce the `@derive` directive
having a similar behavior as `@lookup`. It'll automatically infer the
fields that can be derived from the names & types. `@is` is now only
used to specify the mapping explicitly.

We enforce that at least one `@key` directive must match, the goal is
not to support completely arbitrary data remappings.