chore(gb-9090): remove sdk 0.8, update to wasmtime 32 (#3165)
The issue why we could not update was how we had some types defined that
were not used anywhere in the WIT. Also our 0.8.0 support was so
different and so broken, that I just deleted it. Nothing is compiled
against that right now, it was pre-pre-pre extensions and we can throw
it away.