feat(engine): Parse composite require directive
Adding it to the schema, but not doing anything with it yet.

I've spent a ton of time trying to figure out why gateway tracing tests
weren't working, I'm still not sure... But it seems like we only
export/keep very few spans in `otel_traces` per service name? Was
generating some in schema building, saw some of those, but none of the
actual request for who knows why.