chore: fixes
Simplified the extension serialization. Its primary function is to be a
stable and easily understandable format, so we can quickly identify
breaking change among other things. The current implementation with
bitflags makes me profoundly uneasy.

Replaced direct access to the event queue in the shared context with
`push-event`. The only one we to be able to pop events is the
`on_response` hook.

I've optimized the `push-event` to not read data if not necessary and
added a `should_queue_event` in the internal SDK.

I've also fixed the event queue which wasn't properly retrieved in the
engine as integration-tests had a different  context type.