cli: error in extension commands when config is not found (#3269)
Currently, if the configuration is not found, an empty configuration is
assumed. This leads to commands like `grafbase extension install`
returning with a successful exit status when no configuration is
present. Since the user would expect a failure in this case, the CLI
should return with a non-zero exit status.

This commit makes it happen.

closes GB-9277