gateway(docker/helm): align UIDs between podSecurityContext and image (#3273)
Before this commit:

- The grafbase user inside the container is created without a hardcoded
UID. In practice, it ends up being 1000
- The podSecurityContext in the helm chart makes the gateway run with
user 1001

That leads to a mismatch, with the gateway running as a completely
unpriviledged user, that doesn't have access to any files baked into the
docker image in later steps.

This is in particular a problem when you install and bake gateway
extensions into the docker image. The gateway doesn't have permissions
to read them.

In this commit,

- the grafbase user gets UID 1000 assigned explicitly, so it shouldn't
change in practice for existing deployments (less breaking)
- the helm chart runs containers as user 1000

So the gateway should run as the grafbase user and have access to the
files created in dockerfiles building upon the base gateway dockerfile
from now on.