graphql-composition: fix panic when sets of enums values that should match have different lengths (#3292)
When an enum is used both in input and output positions, composition
enforces that definitions of that enum in all subgraphs must have
exactly the same values.

In the case where one subgraph had more values than another, composition
would do an out of bound access to test equality, and panic. It's order
dependent, so it was not encountered in the existing tests. This commit
fixes the out of bound errors, preserving the same logic, and improves
the composition error message formulation.

closes GB-9354