graphql-composition: improve assertions in declarative tests (#3293)
I was writing a composite schemas spec validation that emits a warning,
and was reminded that our test setup did not allow testing warnings that
are not accompanied by an error.

So this PR reorganizes how and what we assert in composition tests.

Before:

- Either successful composition or failure. Warnings were only displayed
next to the errors in the failure case, not in the success case.
- We could have dangling files, where we assert an error but there is
still an api.graphql file around.
- Diagnostics were rendered as a GraphQL file that was all comments.

After this PR:

- We always assert all the relevant files. The snapshots are just empty
when no diagnostic, or no schema is present. That avoids the dangling
file problem.
- We always render diagnostics.
- The diagnostics aren't GraphQL comments in an empty schema anymore, we
can render them however we please. So now we can render multi-line
diagnostics on multiple lines.

I also was reminded how not-ideal the current diagnostics API is, so I
exposed more from the `diagnostics` module (see changelog entry).