chore(wasm): Rework context & error handling
This huge PR does two things:

1. Rework the ExtensionContext
2. Remove `wasi_componen_loader::{Error, ErroResponse}` and use
  `wasmtime::Result`.

For the first we had the authorization state in the current wasm context
and added the contract key. I also intended to add the authentication
token in there. But I realized that this was just a bad pattern. This
forces us to have some form of mutability we weren't handling batch
operations within a request correctly as each should have their own
authorization state. So now I extracted all of this from the now
`WasmContext` which only holds the `event_queue`. The rest is handled
separately by the engine. We forward the token and contract key in a
`RequestExtensions` struct exposed by the engine and keep the
authorization state in the `GraphqlRequestContext` which is tied to the
operation rather than the request.

Finally, midway I started reworking our error handling in `wasi_componen_loader`.
The core problem today was that we mix everything together in a single `Error` enum.
While it's not necessarily intuitive at first, to keep a double `Result`
I find that it more clearly differentiates wasm errors from the guest
output. From there I've added automatic poisoning handling with a
`wasmsafe!` macro (couldn't figure out the lifetime issue...). The public metadata
endpoint method for authentication wasn't poisoning the instance correctly for example.

I've also done various clean up in the wasi_componen_loader crate.