chore(wasm): Properly share state & Cache by name.
Today, except the cache, nothing is shared between the extension
instances. So fixed that by having a separate `ExtensionState` and
renamed `WasiState` to `InstanceState`.

In addition, I've replaced the current global Cache with one that must
be created explicitly by the extension with a name. For a given name,
the data will be shared across instances. This allows extension
developer to control the size of the cache (which was unlimited before
as it was a `DashMap` rather than a LRU cache) but also have different
caches. Furthermore added control over the timeout used for the `get`
which is hardcoded to 5 seconds today.

We still have the thundering herd problem in case of timeout,
and I'm not really sure how to
best solve it. Without the TTL we could use the `quick_cache`
implementation by providing a guard to the guest I think.