feat(sdk): Add on_subgraph_request hook.
So adding the `Hooks.on_subgraph_request`. While doing it I realized we could handle headers in a better
way. Today we prevent the guest from dropping the headers and take them
back. Honestly I'm not even really why we decided to go that way, but
it's susceptible to memory leaks if we forget to remove them on the host
side. So now I'm treated headers like every other object and the guest
returns them. This will also make it easier to implement the WIT in
other languages as you can't guess in the WIT that you should not hold
onto the headers, it'll end up being kind of a null pointer.

As headers are used in a bunch of places, had to rework a lot of the
wit, and profited from it to move many types to `*-types.wit`. While not
strictly necessary, it's easier to re-use types that way. Maybe one day
we'll merge a bunch of those into a single `types.wit`.

By doing this I think I've prevented a memory leak in the event queue.
for subgraph responses, we were sending a `Arc<http::HeaderMap>` as a
`Lease::Shared`, meaning that it won't ever be dropped by the guest and
we'll keep it indefinitely in the resources table and we weren't
removing it later. Furthermore it would end up as read-only headers in
the guest while it's presented as an owned variant leading to panics for
any mutations. So now all of this should be clearer and we won't have
this kind of problems.