protoc-gen-grafbase-subgraph: support for Query fields and GraphQL directives
Implemented as protobuf options. See the new sections in the README for details.

I had to switch the whole crate to be built on the protobuf crate instead of prost to make this work, since prost does not preserve unknown fields and does not support custom options in any way.

Also release 0.2.0.

closes GB-9362