protoc-gen-grafbase-subgraph: Query fields, GraphQL directives (#3317)
Map RPC methods to Query fields, and add arbitrary GraphQL directives to
the generated schema.

Implemented as protobuf options. See the new sections in the README for
details.

I had to switch the whole crate to be built on the protobuf crate
instead of prost to make this work, since prost does not preserve
unknown fields and does not support custom options in any way.

Also release 0.2.0.

closes GB-9362