graphql-composition: implement composite schemas override validations  (#3374)
We have to disentangle the validations on shareable / external /
override. The composite schemas spec is a pretty good framework for
thinking about these, so let's implement these validations, so we can
progressively unravel the existing ones. We should be more permissive,
but still correct in the end.

The first commit in this PR also simplifies how fields are accessed in
the Subgraphs data structure.