graphql-composition: implement composite schemas rules for @shareable (#3377)
This makes the validation diagnostics better, and the logic clearer.

It also makes the validation more relaxed: if any subgraph defines the
field as shareable, then others don't need to annotate with `@shareable`
as well. It should not make any schema that composes today fail to
compose.

Also, improve the diagnostic message when you try to compose zero
subgraphs.