graphql-composition: speed up enum composition
This delivers a significant performance improvement. We measured a 66% improvement on a very large federated graph.

The substance of the optimization is simple: we iterate over BTrees to compose enums, when checking in which positions they are used. This PR changes these code paths to iterate over flat vectors instead.