graphql-composition: handle entities with type extensions (#3454)
This is important for federation v1. If there are definitions
in-between, subgraph schemas with something like:

```graphql
extend type A {
  fieldC: String
}

# ... definitions in between

type A @key(fields: "a b") {
  a: Int
  b: Int
}
```

would lead to non-contiguous sections of `subgraphs.keys.keys` to have
the same definition id. So when we end up binary searching there, we
would land in either section, and miss the key. This is solved by
sorting the keys after subgraph ingestion.

To avoid the same problems with extra directives and fields in the
future, this PR also makes sure we do this there, and stop relying on
the BTreeMap with fields sorted by name, so the bug would manifest
immediately for fields too if we didn't sort.