graphql-composition: remove walkers (#3467)
This finishes a refactoring I started progressively a while ago. Replace
the Walker pattern:

```graphql
struct Walker<'a, Id> {
  subgraphs: &'a Subgraphs,
  id: Id
}
```

with the View pattern:

```graphql
struct View<'a, Id, Record> {
  id: Id,
  record: &'a Record,
}
```

With methods on the ids and records, and a Deref implementation that
derefs to the record. With the walker pattern, we ended up pushing the
records into the id field in some cases (fields, arguments), to avoid
repeated lookups on each method call.

The walker pattern had pros:

- Enables method chaining for traversing the subgraphs
- The `Subgraphs` argument stays implicit

And cons:

- Excessive lookups when you need the record, or large walkers with
potentially unnecessary records included. You always have the record in
the walker or never, no alternative.
- The methods are not transparent on whether they are just a field
access, a vec dereferencing or something more expensive
- An array of 10 walkers has 10 unnecessary references to Subgraphs.
This is wasted space.

In the end, the view pattern is less code, more explicit, simpler, and
close enough in ergonomics when traversing the schemas.